localisation/requirements-gpu.txt
tensorflow-gpu
opencv-python
lxml
tqdm
absl-py
matplotlib
easydict
pillow
pytesseract